[DIAGNOSTIQUE][SAISIE-CONTEXTE] Ajoute l'entrepôt qui permet de récupérer les informations du diagnostique depuis l'API
