+++ mon-aide-cyber-ui/src/domaine/Entrepots.ts
@@ -1,0 +1,9 @@
+import { EntrepotDiagnostique } from "./diagnostique/Diagnostique.ts";
+
+export interface Entrepot<T> {
+  lis(identifiant: string): Promise<T>;
+}
+
+export interface Entrepots {
+  diagnostique(): EntrepotDiagnostique;
+}
