+++ mon-aide-cyber-ui/src/domaine/diagnostique/Diagnostique.ts
@@ -1,6 +1,9 @@
 import { Referentiel } from "./Referentiel.ts";
 import { Aggregat } from "../Aggregat.ts";
 
+import { Entrepot } from "../Entrepots.ts";
+
 export type Diagnostique = Aggregat & {
   referentiel: Referentiel;
 };
+export type EntrepotDiagnostique = Entrepot<Diagnostique>;
