+++ mon-aide-cyber-ui/src/stories/Diagnostique.stories.tsx
@@ -3,7 +3,7 @@ import { waitFor, within } from "@storybook/testing-library";
 import { ComposantDiagnostique } from "../composants/ComposantDiagnostique.tsx";
 import { expect } from "@storybook/jest";
 import { FournisseurEntrepots } from "../fournisseurs/FournisseurEntrepot.ts";
-import { EntrepotDiagnostiqueMemoire } from "../../test/infrastructure/entrepots/EntrepotMemoire.ts";
+import { EntrepotDiagnostiqueMemoire } from "../../test/infrastructure/entrepots/EntrepotsMemoire.ts";
 import { unDiagnostique } from "../../test/consructeurs/constructeurDiagnostique.ts";
 import { unReferentiel } from "../../test/consructeurs/construceurReferentiel.ts";
 import { faker } from "@faker-js/faker/locale/fr";
